+++ crates/config-loader/src/lib.rs
@@ -1,0 +1,10 @@
+mod config_cache;
+mod config_finder;
+mod config_loader;
+mod extensions_config_ext;
+mod formats;
+mod toolchains_config_ext;
+
+pub use config_loader::*;
+pub use extensions_config_ext::*;
+pub use toolchains_config_ext::*;