+++ crates/config/src/toolchain/mod.rs
@@ -1,7 +1,5 @@
-mod bin_config;
 mod moon_config;
 mod proto_config;
 
-pub use bin_config::*;
 pub use moon_config::*;
 pub use proto_config::*;