+++ crates/config/tests/template_frontmatter_test.rs
@@ -1,6 +1,6 @@
 mod utils;
 
-use moon_config::TemplateFrontmatterConfig;
+use moon_config_loader::ConfigLoader;
 use utils::*;
 
 mod template_frontmatter {
@@ -11,12 +11,16 @@ mod template_frontmatter {
         expected = "unknown field `title`, expected one of `$schema`, `force`, `to`, `skip`"
     )]
     fn error_unknown_field() {
-        test_parse_config("title: test", |code| TemplateFrontmatterConfig::parse(code));
+        test_parse_config("title: test", |code| {
+            ConfigLoader::parse_template_frontmatter_config(code)
+        });
     }
 
     #[test]
     fn loads_defaults() {
-        let config = test_parse_config("", |code| TemplateFrontmatterConfig::parse(code));
+        let config = test_parse_config("", |code| {
+            ConfigLoader::parse_template_frontmatter_config(code)
+        });
 
         assert!(!config.force);
         assert!(!config.skip);
@@ -26,7 +30,7 @@ mod template_frontmatter {
     #[test]
     fn can_set_force_skip() {
         let config = test_parse_config("force: true\nskip: true", |code| {
-            TemplateFrontmatterConfig::parse(code)
+            ConfigLoader::parse_template_frontmatter_config(code)
         });
 
         assert!(config.force);
@@ -36,19 +40,23 @@ mod template_frontmatter {
     #[test]
     #[should_panic(expected = "invalid type: integer `123`, expected a boolean")]
     fn invalid_force() {
-        test_parse_config("force: 123", |code| TemplateFrontmatterConfig::parse(code));
+        test_parse_config("force: 123", |code| {
+            ConfigLoader::parse_template_frontmatter_config(code)
+        });
     }
 
     #[test]
     #[should_panic(expected = "invalid type: string \"abc\", expected a boolean")]
     fn invalid_skip() {
-        test_parse_config("skip: abc", |code| TemplateFrontmatterConfig::parse(code));
+        test_parse_config("skip: abc", |code| {
+            ConfigLoader::parse_template_frontmatter_config(code)
+        });
     }
 
     #[test]
     fn can_set_to() {
         let config = test_parse_config("to: some/path", |code| {
-            TemplateFrontmatterConfig::parse(code)
+            ConfigLoader::parse_template_frontmatter_config(code)
         });
 
         assert_eq!(config.to, Some("some/path".into()));
@@ -57,6 +65,8 @@ mod template_frontmatter {
     #[test]
     #[should_panic(expected = "invalid type: boolean `true`, expected a string")]
     fn invalid_to() {
-        test_parse_config("to: true", |code| TemplateFrontmatterConfig::parse(code));
+        test_parse_config("to: true", |code| {
+            ConfigLoader::parse_template_frontmatter_config(code)
+        });
     }
 }